refactor: organize renderer and source modules
