+++ src/renderers/graphical/gutter.rs
