+++ src/renderers/graphical/report.rs
