+++ src/renderers/graphical/span.rs
