+++ src/renderers/mod.rs
@@ -1,0 +1,7 @@
+//! Diagnostic renderers included with `miette`.
+
+pub use graphical::*;
+pub use json::*;
+
+mod graphical;
+mod json;
